EduGuideBot_FINAL/requirements.txt
python-telegram-bot[ext]
python-dotenv
pandas
scikit-learn
joblib 